+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 - (Documentation) ManualUpgrade Docs
 - (Documentation) Add Required & Skip in Docs
 - (Feature) (Platform) ECS Storage
-- (Bugfix) (Platform) Prevent NPE in case of missing Helm Release 
+- (Bugfix) (Platform) Prevent NPE in case of missing Helm Release
+- (Feature) Compact Action
 
 ## [1.2.50](https://github.com/arangodb/kube-arangodb/tree/1.2.50) (2025-07-04)
 - (Feature) (Platform) MetaV1 Integration Service